data/requirements.txt
requests==2.32.3
openai==1.68.2
python-dotenv==1.1.0
firebase-admin==6.7.0